refactor: split ui update module
