--- src/egui_app/ui/update.rs
@@ -1,494 +1,0 @@
-use crate::egui_app::state::{FocusContext, TriageFlagColumn};
-use crate::sample_sources::SampleTag;
-use crate::selection::SelectionEdge;
-use eframe::egui;
-
-use super::EguiApp;
-use super::input::{InputSnapshot, copy_shortcut_pressed};
-#[cfg(target_os = "windows")]
-use super::platform;
-
-struct FocusFlags {
-    browser: bool,
-    folder: bool,
-    waveform: bool,
-    sources: bool,
-    collection_sample: bool,
-    collections_list: bool,
-}
-
-impl FocusFlags {
-    fn from_context(context: FocusContext) -> Self {
-        Self {
-            browser: matches!(context, FocusContext::SampleBrowser | FocusContext::None),
-            folder: matches!(context, FocusContext::SourceFolders),
-            waveform: matches!(context, FocusContext::Waveform),
-            sources: matches!(context, FocusContext::SourcesList),
-            collection_sample: matches!(context, FocusContext::CollectionSample),
-            collections_list: matches!(context, FocusContext::CollectionsList),
-        }
-    }
-}
-
-impl eframe::App for EguiApp {
-    fn update(&mut self, ctx: &egui::Context, frame: &mut eframe::Frame) {
-        self.prepare_frame(ctx, frame);
-        let focus_context = self.controller.ui.focus.context;
-        let focus_flags = FocusFlags::from_context(focus_context);
-        self.handle_focus_side_effects(&focus_flags);
-        let input = InputSnapshot::capture(ctx);
-        let feedback_modal_open = self.controller.ui.feedback_issue.open;
-        if !feedback_modal_open {
-            self.handle_space_shortcut(ctx, &input);
-            self.handle_copy_shortcut(ctx);
-            self.handle_escape_shortcut(ctx, &input);
-            self.handle_window_shortcuts(ctx);
-            self.handle_arrow_keys(ctx, &focus_flags, &input);
-            self.process_hotkeys(ctx, focus_context);
-        }
-        self.render_ui(ctx, &input, focus_context);
-    }
-}
-
-impl EguiApp {
-    fn prepare_frame(&mut self, ctx: &egui::Context, _frame: &mut eframe::Frame) {
-        self.apply_visuals(ctx);
-        self.ensure_initial_focus(ctx);
-        let feedback_modal_open = self.controller.ui.feedback_issue.open;
-        #[cfg(target_os = "windows")]
-        self.controller
-            .set_drag_hwnd(platform::hwnd_from_frame(_frame));
-        #[cfg(target_os = "windows")]
-        if !feedback_modal_open {
-            let pixels_per_point = ctx.pixels_per_point();
-            self.controller.ui.drag.os_cursor_pos = platform::hwnd_from_frame(_frame)
-                .and_then(|hwnd| platform::cursor_pos_in_client_points(hwnd, pixels_per_point));
-            let left_mouse_down = platform::left_mouse_button_down();
-            self.controller
-                .ui
-                .drag
-                .update_os_mouse_state(left_mouse_down);
-        }
-        self.controller.tick_playhead();
-        if !feedback_modal_open {
-            if let Some(pos) = ctx.input(|i| i.pointer.hover_pos().or_else(|| i.pointer.interact_pos()))
-            {
-                let shift_down = ctx.input(|i| i.modifiers.shift);
-                self.controller.refresh_drag_position(pos, shift_down);
-            }
-        } else {
-            self.controller.ui.drag.position = None;
-            self.controller.ui.drag.label.clear();
-            self.controller.ui.drag.clear_all_targets();
-        }
-        #[cfg(target_os = "windows")]
-        if !feedback_modal_open {
-            if self.controller.ui.drag.payload.is_some() {
-                ctx.request_repaint();
-            }
-            let window_inside = self
-                .controller
-                .ui
-                .drag
-                .payload
-                .is_some()
-                .then(|| platform::hwnd_from_frame(_frame))
-                .flatten()
-                .and_then(platform::cursor_inside_hwnd);
-
-            if self.controller.ui.drag.payload.is_some()
-                && matches!(window_inside, Some(false))
-                && !self.controller.ui.drag.external_started
-            {
-                // The pointer has left the window during an in-app drag. Hide all internal drag
-                // visuals (ghost + hover highlights) so the only remaining user action is to drop
-                // externally once the OS drag-out is triggered.
-                self.controller.ui.drag.position = None;
-                self.controller.ui.drag.label.clear();
-                self.controller.ui.drag.clear_all_targets();
-            }
-
-            let (pointer_outside, pointer_left) = ctx.input(|i| {
-                if self.controller.ui.drag.payload.is_some() {
-                    if matches!(window_inside, Some(false)) {
-                        // Once the pointer leaves the window during a drag, permanently disable
-                        // in-app dragging until the gesture completes (external-only).
-                        self.controller.ui.drag.pointer_left_window = true;
-                    }
-                    let pointer_gone = i
-                        .events
-                        .iter()
-                        .any(|e| matches!(e, egui::Event::PointerGone));
-                    if pointer_gone {
-                        self.controller.ui.drag.pointer_left_window = true;
-                    }
-                } else {
-                    self.controller.ui.drag.pointer_left_window = false;
-                }
-
-                // Prefer the OS cursor/window geometry check when available; it's robust even
-                // when egui stops reporting pointer positions outside the window.
-                let inside = window_inside.unwrap_or_else(|| {
-                    let hover_pos = i.pointer.hover_pos();
-                    let interact_pos = i.pointer.interact_pos();
-                    hover_pos.is_some()
-                        || (interact_pos.is_some() && !self.controller.ui.drag.pointer_left_window)
-                });
-                let outside = self.controller.ui.drag.payload.is_some() && !inside;
-                let left = match window_inside {
-                    Some(true) => false,
-                    Some(false) => true,
-                    None => self.controller.ui.drag.pointer_left_window,
-                };
-                (outside, left)
-            });
-            self.controller
-                .maybe_launch_external_drag(pointer_outside, pointer_left);
-        }
-    }
-
-    fn handle_focus_side_effects(&mut self, focus: &FocusFlags) {
-        if !focus.browser && !focus.waveform {
-            self.controller.blur_browser_focus();
-        }
-        if focus.collection_sample {
-            self.controller.ui.browser.autoscroll = false;
-        }
-    }
-
-    fn handle_space_shortcut(&mut self, ctx: &egui::Context, input: &InputSnapshot) {
-        if !input.space {
-            return;
-        }
-        if ctx.wants_keyboard_input() {
-            return;
-        }
-        let ctrl_or_command = input.ctrl_or_command();
-        if input.shift {
-            let handled = self.controller.replay_from_last_start();
-            if !handled {
-                self.controller.toggle_play_pause();
-            }
-        } else if ctrl_or_command {
-            let handled = self.controller.play_from_cursor();
-            if !handled {
-                self.controller.toggle_play_pause();
-            }
-        } else {
-            self.controller.toggle_play_pause();
-        }
-        consume_keypress(ctx, input, egui::Key::Space);
-    }
-
-    fn handle_copy_shortcut(&mut self, ctx: &egui::Context) {
-        if copy_shortcut_pressed(ctx) {
-            self.controller.copy_selection_to_clipboard();
-        }
-    }
-
-    fn handle_escape_shortcut(&mut self, ctx: &egui::Context, input: &InputSnapshot) {
-        if !input.escape {
-            return;
-        }
-        if self.controller.ui.progress.visible {
-            self.controller.request_progress_cancel();
-        }
-        self.controller.handle_escape();
-        if self.controller.ui.hotkeys.overlay_visible {
-            self.controller.ui.hotkeys.overlay_visible = false;
-            ctx.input_mut(|state| state.consume_key(egui::Modifiers::default(), egui::Key::Escape));
-        }
-        ctx.input_mut(|state| state.consume_key(egui::Modifiers::default(), egui::Key::Escape));
-    }
-
-    fn handle_window_shortcuts(&self, ctx: &egui::Context) {
-        if let Some(new_maximized) = ctx.input(|i| {
-            if i.key_pressed(egui::Key::F11) {
-                Some(!i.viewport().maximized.unwrap_or(false))
-            } else {
-                None
-            }
-        }) {
-            ctx.send_viewport_cmd(egui::ViewportCommand::Maximized(new_maximized));
-        }
-    }
-
-    fn handle_arrow_keys(
-        &mut self,
-        ctx: &egui::Context,
-        focus: &FocusFlags,
-        input: &InputSnapshot,
-    ) {
-        if ctx.wants_keyboard_input() {
-            return;
-        }
-        let browser_has_selection = self.controller.ui.browser.selected.is_some();
-        let ctrl_or_command = input.ctrl_or_command();
-        self.handle_arrow_down(ctx, focus, input);
-        self.handle_arrow_up(ctx, focus, input);
-        self.handle_arrow_right(ctx, focus, input, browser_has_selection, ctrl_or_command);
-        self.handle_arrow_left(ctx, focus, input, browser_has_selection, ctrl_or_command);
-    }
-
-    fn handle_arrow_down(
-        &mut self,
-        ctx: &egui::Context,
-        focus: &FocusFlags,
-        input: &InputSnapshot,
-    ) {
-        if !input.arrow_down {
-            return;
-        }
-        if focus.collection_sample {
-            self.controller.nudge_collection_sample(1);
-        } else if focus.browser {
-            if self.controller.random_navigation_mode_enabled() {
-                self.controller.play_random_visible_sample();
-            } else if input.shift {
-                self.controller.grow_selection(1);
-            } else {
-                self.controller.nudge_selection(1);
-            }
-        } else if focus.folder {
-            self.controller.nudge_folder_selection(1, input.shift);
-        } else if focus.waveform {
-            self.controller.zoom_waveform(false);
-        } else if focus.sources {
-            self.controller.nudge_source_selection(1);
-        } else if focus.collections_list {
-            self.controller.nudge_collection_row(1);
-        }
-        consume_keypress(ctx, input, egui::Key::ArrowDown);
-    }
-
-    fn handle_arrow_up(&mut self, ctx: &egui::Context, focus: &FocusFlags, input: &InputSnapshot) {
-        if !input.arrow_up {
-            return;
-        }
-        if focus.collection_sample {
-            self.controller.nudge_collection_sample(-1);
-        } else if focus.browser {
-            if self.controller.random_navigation_mode_enabled() {
-                self.controller.play_previous_random_sample();
-            } else if input.shift {
-                self.controller.grow_selection(-1);
-            } else {
-                self.controller.nudge_selection(-1);
-            }
-        } else if focus.folder {
-            self.controller.nudge_folder_selection(-1, input.shift);
-        } else if focus.waveform {
-            self.controller.zoom_waveform(true);
-        } else if focus.sources {
-            self.controller.nudge_source_selection(-1);
-        } else if focus.collections_list {
-            self.controller.nudge_collection_row(-1);
-        }
-        consume_keypress(ctx, input, egui::Key::ArrowUp);
-    }
-
-    fn handle_arrow_right(
-        &mut self,
-        ctx: &egui::Context,
-        focus: &FocusFlags,
-        input: &InputSnapshot,
-        browser_has_selection: bool,
-        ctrl_or_command: bool,
-    ) {
-        if !input.arrow_right {
-            return;
-        }
-        if focus.waveform {
-            self.handle_waveform_arrow(input, ctrl_or_command, true);
-        } else if focus.folder {
-            self.controller.expand_focused_folder();
-        } else if ctrl_or_command && focus.browser && browser_has_selection {
-            self.controller.move_selection_column(1);
-        } else if focus.collection_sample {
-            self.controller
-                .tag_selected_collection_sample(SampleTag::Keep);
-        } else if focus.browser && browser_has_selection {
-            let col = self.controller.ui.browser.selected.map(|t| t.column);
-            let target = if matches!(col, Some(TriageFlagColumn::Trash)) {
-                SampleTag::Neutral
-            } else {
-                SampleTag::Keep
-            };
-            self.controller.tag_selected(target);
-        }
-        consume_keypress(ctx, input, egui::Key::ArrowRight);
-    }
-
-    fn handle_arrow_left(
-        &mut self,
-        ctx: &egui::Context,
-        focus: &FocusFlags,
-        input: &InputSnapshot,
-        browser_has_selection: bool,
-        ctrl_or_command: bool,
-    ) {
-        if !input.arrow_left {
-            return;
-        }
-        if focus.waveform {
-            self.handle_waveform_arrow(input, ctrl_or_command, false);
-        } else if focus.folder {
-            self.controller.collapse_focused_folder();
-        } else if ctrl_or_command && focus.browser && browser_has_selection {
-            self.controller.move_selection_column(-1);
-        } else if focus.collection_sample {
-            self.controller.tag_selected_collection_left();
-        } else if focus.browser && browser_has_selection {
-            self.controller.tag_selected_left();
-        }
-        consume_keypress(ctx, input, egui::Key::ArrowLeft);
-    }
-
-    fn handle_waveform_arrow(
-        &mut self,
-        input: &InputSnapshot,
-        ctrl_or_command: bool,
-        move_right: bool,
-    ) {
-        let was_playing = self.controller.is_playing();
-        let has_selection = self.controller.ui.waveform.selection.is_some();
-        if input.shift {
-            self.adjust_waveform_selection(
-                input,
-                ctrl_or_command,
-                move_right,
-                was_playing,
-                has_selection,
-            );
-            return;
-        }
-        self.move_waveform_playhead(input, move_right, was_playing);
-    }
-
-    fn adjust_waveform_selection(
-        &mut self,
-        input: &InputSnapshot,
-        ctrl_or_command: bool,
-        move_right: bool,
-        was_playing: bool,
-        has_selection: bool,
-    ) {
-        if ctrl_or_command {
-            self.handle_waveform_chord_selection(input, move_right, was_playing, has_selection);
-            return;
-        }
-        self.handle_waveform_shift_selection(input, move_right, was_playing, has_selection);
-    }
-
-    fn handle_waveform_chord_selection(
-        &mut self,
-        input: &InputSnapshot,
-        move_right: bool,
-        was_playing: bool,
-        has_selection: bool,
-    ) {
-        let start_direction_is_left = !move_right;
-        if has_selection {
-            self.controller.nudge_selection_edge(
-                SelectionEdge::Start,
-                start_direction_is_left,
-                input.alt,
-            );
-        } else {
-            self.controller.create_selection_from_playhead(
-                start_direction_is_left,
-                was_playing,
-                input.alt,
-            );
-        }
-    }
-
-    fn handle_waveform_shift_selection(
-        &mut self,
-        input: &InputSnapshot,
-        move_right: bool,
-        was_playing: bool,
-        has_selection: bool,
-    ) {
-        let start_direction_is_left = !move_right;
-        if has_selection {
-            self.controller
-                .nudge_selection_edge(SelectionEdge::End, move_right, input.alt);
-            return;
-        }
-        self.controller.create_selection_from_playhead(
-            start_direction_is_left,
-            was_playing,
-            input.alt,
-        );
-    }
-
-    fn move_waveform_playhead(
-        &mut self,
-        input: &InputSnapshot,
-        move_right: bool,
-        was_playing: bool,
-    ) {
-        let step = if move_right { 1 } else { -1 };
-        self.controller
-            .move_playhead_steps(step, input.alt, was_playing);
-    }
-}
-
-fn consume_keypress(ctx: &egui::Context, input: &InputSnapshot, key: egui::Key) {
-    let mut modifiers = egui::Modifiers::default();
-    modifiers.shift = input.shift;
-    modifiers.alt = input.alt;
-    modifiers.ctrl = input.ctrl;
-    modifiers.command = input.command;
-    ctx.input_mut(|state| state.consume_key(modifiers, key));
-}
-
-#[cfg(test)]
-mod tests {
-    use super::InputSnapshot;
-    use crate::egui_app::controller::EguiController;
-    use crate::egui_app::ui::EguiApp;
-    use crate::egui_app::ui::hotkey_runtime::KeyFeedback;
-    use crate::waveform::WaveformRenderer;
-    use eframe::egui;
-
-    fn test_app() -> EguiApp {
-        let renderer = WaveformRenderer::new(8, 8);
-        let controller = EguiController::new(renderer, None);
-        EguiApp {
-            controller,
-            visuals_set: false,
-            waveform_tex: None,
-            last_viewport_log: None,
-            sources_panel_rect: None,
-            sources_panel_drop_hovered: false,
-            sources_panel_drop_armed: false,
-            selection_edge_offset: None,
-            selection_slide: None,
-            pending_chord: None,
-            key_feedback: KeyFeedback::default(),
-            requested_initial_focus: false,
-        }
-    }
-
-    #[test]
-    fn ctrl_shift_space_prefers_replay_from_last_start_over_cursor() {
-        let ctx = egui::Context::default();
-        let mut app = test_app();
-        app.controller.ui.waveform.last_start_marker = Some(0.8);
-        app.controller.ui.waveform.cursor = Some(0.2);
-
-        app.handle_space_shortcut(
-            &ctx,
-            &InputSnapshot {
-                space: true,
-                ctrl: true,
-                shift: true,
-                ..InputSnapshot::default()
-            },
-        );
-
-        assert_eq!(app.controller.ui.waveform.last_start_marker, Some(0.8));
-    }
-}
